# swagger/openapi.yaml
openapi: 3.0.0
info:
  title: CityLink
  version: 0.0.1
servers:
  - url: http://localhost:4000/api
    description: Dev server
paths:
...
  /routes/getRouteRecords:
    get:
      summary: Получение списка доступных записанных маршрутов
      tags: 
        - Routes
      operationId: getRouteRecords
      parameters:
        - name: departureCity
          in: query
          schema:
            type: string
            description: Город отправления
            example: Ставрополь   
        - name: destinationCity
          in: query
          schema:
            type: string
            description: Город прибытия
            example: Пятигорск
        - name: departureDate
          in: query
          schema:
            type: string
            description: Время отправления
            example: 1704206143000(timestamp)
    
      responses:
        '200':
          description: Успешное получение списка записанных маршрутов
          content:
            'application/json':
              schema:
                type: object
                properties:
                  routes:
                    description: Массив объектов
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: number
                          description: id записи маршрута
                          example: 82
                        price:
                          type: number
                          description: Цена билета
                          example: 500
                        departureDate: 
                          type: string
                          description: Дата отправления
                          example: 2023-12-21T08:00:00.000Z
                        duration:
                          type: number
                          description: Время в пути
                          example: 5
                        departureCity:
                          type: string
                          description: Город отправления
                          example: Ставрополь
                        destinationCity:
                          type: string
                          description: Город прибытия
                          example: Пятигорск
                        
        '400':
          description:  Ошибка в отправляемых данных
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/400"
                
        '500':
          description:  Ошибка на стороне сервера
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/500" 
  /routes/getRoutes:
    get:
      summary: Получение списка маршрутов
      tags: 
        - Routes
      operationId: getRoutes
      parameters:
        - name: departureCity
          in: query
          schema:
            type: string
            description: Город отправления
            example: Ставрополь   
        - name: destinationCity
          in: query
          schema:
            type: string
            description: Город прибытия
            example: Пятигорск
        - name: Authorization
          in: header
          schema:
            type: string
            description: Токен авторизации
            example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImViMzhmYzFkLTI4YWQtNDY1Mi05OWU5LWJjNjA5YTBkZjllYSIsImlhdCI6MTcwNTg0MzYzOSwiZXhwIjoxNzA2Mjc1NjM5fQ.Xm-cygFpoVrHMaNoiXamGnDb7_LDbGPubNnDWK6BWow
    
      responses:
        '200':
          description: Успешное получение списка маршрутов
          content:
            'application/json':
              schema:
                type: object
                properties:
                  routes:
                    description: Массив объектов
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: number
                          description: id маршрута
                          example: 82
                        duration:
                          type: number
                          description: Время в пути
                          example: 5
                        departureCity:
                          type: string
                          description: Город отправления
                          example: Ставрополь
                        destinationCity:
                          type: string
                          description: Город прибытия
                          example: Пятигорск
                        
        '400':
          description:  Ошибка в отправляемых данных
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/400"
                
        '500':
          description:  Ошибка на стороне сервера
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/500" 
  /routes/getCities:
    get:
      summary: Получение городов по названию
      tags: 
        - Routes
      operationId: getCities
      parameters:
        - name: cityName
          in: query
          schema:
            type: string
            description: Полное или часть названия города
            example: Ста  
        - name: Authorization
          in: header
          schema:
            type: string
            description: Токен авторизации
            example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImViMzhmYzFkLTI4YWQtNDY1Mi05OWU5LWJjNjA5YTBkZjllYSIsImlhdCI6MTcwNTg0MzYzOSwiZXhwIjoxNzA2Mjc1NjM5fQ.Xm-cygFpoVrHMaNoiXamGnDb7_LDbGPubNnDWK6BWow
      responses:
        '200':
          description: Успешное получение списка городов
          content:
            'application/json':
              schema:
                type: object
                properties:
                  routes:
                    description: Массив объектов
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: number
                          description: id города
                          example: 1
                        cityName:
                          type: string
                          description: Название города
                          example: Ставрополь
                        
        '400':
          description:  Ошибка в отправляемых данных
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/400"
                
        '500':
          description:  Ошибка на стороне сервера
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/500" 
  /routes/addRoute:
    post:
      summary: Добавление маршрута
      tags: 
        - Routes
      operationId: addRoute
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
            description: Токен авторизации
            example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImViMzhmYzFkLTI4YWQtNDY1Mi05OWU5LWJjNjA5YTBkZjllYSIsImlhdCI6MTcwNTg0MzYzOSwiZXhwIjoxNzA2Mjc1NjM5fQ.Xm-cygFpoVrHMaNoiXamGnDb7_LDbGPubNnDWK6BWow
            
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                departureCity:
                  description: Город отправления
                  type: string
                  example: Ставрополь
                destinationCity:
                  description: Город прибытия
                  type: string
                  example: Пятигорск
                duration:
                  description: Время в пути
                  type: number
                  example: 4
              required:
                - departureCity
                - destinationCity
                - duration
      responses:
        '200':
          description: Успешное добавление города
          content:
            'application/json':
              schema:
                type: object
                properties:
                  id:
                    description: id маршрута
                    type: number
                    example: 10
                  duration:
                    description: Время в пути
                    type: number
                    example: 5
                  departureCityId:
                    description: id города отправления
                    type: number
                    example: 4
                  destinationCityId:
                    description: id города прибытия
                    type: number
                    example: 7
                        
        '400':
          description:  Ошибка в отправляемых данных
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/400"
                
        '500':
          description:  Ошибка на стороне сервера
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/500" 
  /routes/addRouteRecord:
    post:
      summary: Добавление записанного маршрута
      tags: 
        - Routes
      operationId: addRouteRecord
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
            description: Токен авторизации
            example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImViMzhmYzFkLTI4YWQtNDY1Mi05OWU5LWJjNjA5YTBkZjllYSIsImlhdCI6MTcwNTg0MzYzOSwiZXhwIjoxNzA2Mjc1NjM5fQ.Xm-cygFpoVrHMaNoiXamGnDb7_LDbGPubNnDWK6BWow
            
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                departureCity:
                  description: Город отправления
                  type: string
                  example: Ставрополь
                destinationCity:
                  description: Город прибытия
                  type: string
                  example: Пятигорск
                regNumber:
                  description: Регистрационный номер автобуса
                  type: string
                  example: д325мв
                departureDate:
                  description: Время отправления (timestamp)
                  type: number
                  example: 1703361600000
                price:
                  description: Цена билета
                  type: number
                  example: 500
              required:
                - departureCity
                - destinationCity
                - regNumber
                - departureDate
                - price
      responses:
        '200':
          description: Успешное добавление записанного маршрута
          content:
            'application/json':
              schema:
                type: object
                properties:
                  result:
                    type: object
                    properties:
                      id:
                        description: id записанного маршрута
                        type: number
                        example: 10
                      price:
                        description: Цена билета
                        type: number
                        example: 500
                      departureDate:
                        description: Время отправления
                        type: string
                        example: 2023-12-23T20:00:05.000Z
                      routeId:
                        description: id маршрута
                        type: number
                        example: 7
                      busId:
                        description: id автобуса
                        type: number
                        example: 4
                        
                        
        '400':
          description:  Ошибка в отправляемых данных
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/400"
                
        '500':
          description:  Ошибка на стороне сервера
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/500"
...
  /routes/addCity:
    post:
      summary: Добавление города
      tags: 
        - Routes
      operationId: addCity
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
            description: Токен авторизации
            example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImViMzhmYzFkLTI4YWQtNDY1Mi05OWU5LWJjNjA5YTBkZjllYSIsImlhdCI6MTcwNTg0MzYzOSwiZXhwIjoxNzA2Mjc1NjM5fQ.Xm-cygFpoVrHMaNoiXamGnDb7_LDbGPubNnDWK6BWow
            
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                cityName:
                  description: Название города
                  type: string
                  example: Ставрополь
              required:
                - cityName
      responses:
        '200':
          description: Успешное добавление города
          content:
            'application/json':
              schema:
                type: object
                properties:
                  result:
                    type: object
                    properties:
                      id:
                        description: id города
                        type: number
                        example: 10
                      cityName:
                        description: Название города
                        type: string
                        example: Ставрополь
                        
        '400':
          description:  Ошибка в отправляемых данных
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/400"
                
        '500':
          description:  Ошибка на стороне сервера
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/500"
...
components:
  schemas:
    400: 
      description:  Ошибка в отправляемых данных
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Пример сообщения
    500:
      description:  Ошибка на стороне сервера
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Пример сообщения
...